requirements.txt
kafka-python==2.0.2
requests~=2.32.3
pyspark~=3.5.2
cassandra-driver~=3.29.2